Add a workaround for PowerA Wireless Controller
Add a workaround for PowerA Wireless Controller, which doesn't
report valid VID/PID via Bluetooth making it hard to properly tag it
with "uaccess" via UDEV.

The approach taken by this commit is to trigger on input device,
instead of the usual hidraw, and then rely on "udevadm test-builtin"
to apply "uacess" permission at runtime.

The reason to trigger on input device is because it is the first
device in device hierarchy that gets access to reported "name" which
is unique to this particular type of a controller and can be reliably
matched against.